590273 - UPD requiring files and exe/hix class call
