+++ .rubocop_todo.yml
@@ -1,12 +1,19 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config --auto-gen-only-exclude --exclude-limit 10000`
-# on 2022-10-13 09:47:53 UTC using RuboCop version 1.36.0.
+# on 2022-10-18 09:27:56 UTC using RuboCop version 1.36.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
 # versions of RuboCop, may require this file to be generated again.
 
 # Offense count: 3
+Lint/MissingSuper:
+  Exclude:
+    - 'lib/hix/exe/config.rb'
+    - 'lib/hix/exe/login.rb'
+    - 'lib/hix/exe/new.rb'
+
+# Offense count: 2
 # Configuration parameters: AllowedMethods, AllowedPatterns, IgnoredMethods, CountRepeatedAttributes, Max.
 Metrics/AbcSize:
   Exclude:
